deployment updates

+++ app/models/image.rb
@@ -28,6 +28,6 @@ class Image < ApplicationRecord
   end
 
   def delete_file
-    File.delete(file_path)
+    File.delete(file_path) if File.exist?(file_path)
   end
 end
